requirements.txt
scikit-learn==1.7.2
pandas
joblib
pyserial
fastapi
uvicorn
sqlalchemy
python-multipart
requests